# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tindarr API
  version: 1.0.0-draft
  summary: Self-hosted server behind the Tindarr mobile app and web console.
  description: |
    Contract between a Tindarr server and its two clients: the mobile app and the
    web console the server serves under `/`. Written before the implementation; CI
    validates the server's responses against this file and rejects routes it does
    not document.

    Rules (see docs/adr/0006):
    - Additive changes only within v1. Clients ignore unknown fields and treat
      unknown enum values as "other".
    - Errors are `application/problem+json` with a stable `code`.
    - Slow work answers `202` with `retry_after_ms`; the client polls.
    - Every `429` carries `retry_after_ms` and a `Retry-After` header.
    - CORS is disabled: the console is same-origin.

    Every request (docs/auth.md, sections 1 and 2):
    - The `Host` header must be an IP literal, `localhost`, a name in
      `TINDARR_ALLOWED_HOSTS` or the host of `public_url`; anything else gets
      `400` `host_not_allowed` before routing, on every path but `/healthz`.
    - The client IP and scheme come from `X-Forwarded-For` and
      `X-Forwarded-Proto` only when the peer is in `TINDARR_TRUSTED_PROXIES`.
      The server's origin is `<scheme>://<Host>`.
    - Codes, handles and tokens travel in bodies, cookies or headers, never in a
      path or query string.

    Authentication (docs/adr/0009, docs/auth.md section 2), in this order:
...

...
  license:
...
servers:
...
security:
  - bearer: []
  - cookieSession: []
    csrfToken: []
...
paths:
...
  /api/v1/me:
    get:
      tags: [me]
      summary: The signed-in user
      operationId: getMe
      security:
        - bearer: []
        - cookieSession: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [me, console]
      summary: Delete all of the caller's Tindarr data and sign out everywhere
      description: |
        Votes, profile, preferences and sessions. Nothing is deleted on the media
        server or the request backend. Console only: a phone's token cannot delete
        the account.
      operationId: deleteMyData
      security:
        - cookieSession: []
          csrfToken: []
      parameters:
        - name: confirm
          in: query
          required: true
          schema: { type: string, const: delete-my-data }
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`last_admin`: promote another admin first."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/me/sessions:
    get:
      tags: [me]
      summary: The caller's sessions (one per device)
      operationId: listMySessions
      security:
        - bearer: []
        - cookieSession: []
      responses:
        "200":
          description: Sessions, most recent first.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/me/sessions/{session_id}:
    delete:
      tags: [me, console]
      summary: Revoke one of the caller's sessions
      description: Console only. The app signs itself out with `POST /auth/logout`.
      operationId: revokeMySession
      security:
        - cookieSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: Revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- pairing
...
  /api/v1/admin/users:
    get:
      tags: [admin, console]
      security:
        - cookieSession: []
      summary: Users who signed in at least once
      operationId: listUsers
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/AdminUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
  /api/v1/admin/users/{user_id}:
    patch:
      tags: [admin, console]
      security:
        - cookieSession: []
          csrfToken: []
      summary: Enable, disable, promote or set limits for a user
      description: |
        Disabling a user revokes their sessions. `role` edits the Tindarr side
        only (docs/adr/0010): `admin` sets `promoted`; `user` clears `promoted` and
        clears `media_server_admin` until the user's next sign-in, where the media
        server's flag is read again (a media server administrator is then admin
        again). Only a media server administrator can demote or disable another
        one (`forbidden`). The last enabled admin cannot be demoted or disabled
        (`last_admin`). Disabling sets `disabled_reason` = `admin`.
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                enabled: { type: boolean }
                role: { $ref: "#/components/schemas/Role" }
                daily_generation_limit:
                  type: [integer, "null"]
                  minimum: 0
                  description: Null uses the server default.
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`forbidden` (only a media server administrator can demote or disable another one), `admin_required`, `csrf_failed` or `remote_access_denied`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`last_admin`: the last enabled admin cannot be demoted or disabled."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/admin/users/{user_id}/sessions:
    delete:
      tags: [admin, console]
      security:
        - cookieSession: []
          csrfToken: []
      summary: Sign a user out of every device
      operationId: revokeUserSessions
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: Revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from an app sign-in, pairing or refresh, valid 15 minutes. Mobile app only.
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_session
      description: |
        Web console session (`kind` = `web`): opaque 256-bit token, stored hashed,
        `HttpOnly`, `Secure`, `SameSite=Strict`, `Path=/`, no `Domain`. Expires
        after 24 h idle or 7 days after sign-in. Named `tindarr_session`, without
        `Secure`, only on requests from a private client IP when
        `TINDARR_ALLOW_HTTP_CONSOLE=true` (docs/auth.md, section 2).
    setupSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_setup
      description: |
        Setup session from `POST /setup/claim` (`kind` = `setup`, no user,
        30 minutes absolute, same attributes as the web session cookie; named
        `tindarr_setup` under the plain-HTTP opt-in). Only one is valid at a time.
        Accepted by the setup operations, `GET /auth/web/session`, the web sign-ins
        while setup is pending, and Plex PINs of purpose `owner_token`. Never
        accepted where a web session is expected, and the reverse.
    preAuth:
      type: apiKey
      in: cookie
      name: __Host-tindarr_preauth
      description: |
        Not a session: a random value that binds the console's Plex PIN and Quick
        Connect sign-in handles to the browser that started them (docs/auth.md,
        section 5). Set by `POST /auth/plex/pins` and `POST /auth/quick-connect`
        when called by the console without `code_challenge`, if absent; 15 minutes;
        cleared by a successful web sign-in. Named `tindarr_preauth` under the
        plain-HTTP opt-in.
    csrfToken:
      type: apiKey
      in: header
...
  parameters:
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string }
    UserId:
      name: user_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    BadRequest:
      description: |
        `validation_error` (with `errors`; input values are never echoed),
        `bad_request` (malformed body), or `host_not_allowed` (the `Host` header is
        not an allowed host; any path but `/healthz`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        `unauthorized`: no credential this operation accepts (an `Authorization`
        header on an operation without `bearer`, or a cookie of another kind,
        counts as none), a revoked, idle-expired or past-lifetime session, a
        disabled or unlinked user. `token_expired`: the bearer access token is past
        `exp`; refresh (single-flight) and retry.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: |
        `csrf_failed` (cookie request with an unsafe method: wrong or missing
        `X-CSRF-Token` or `Origin`) or `remote_access_denied` (the user's media
        server policy forbids remote access and the client IP is not private).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    AdminForbidden:
      description: "`admin_required` (the effective role is not `admin`), `csrf_failed` or `remote_access_denied`."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: "`not_found`, also for another user's resource."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: |
            Stable identifier. Known values:
            - request and credentials: validation_error, bad_request,
              host_not_allowed, unauthorized, token_expired, refresh_token_reused,
              csrf_failed, https_required, forbidden, admin_required,
              media_server_admin_required, reauth_required, remote_access_denied,
              not_found, method_not_allowed, rate_limited, internal_error;
            - setup: setup_required, setup_completed, invalid_setup_code,
              setup_session_required, setting_locked;
            - sign-in: invalid_credentials, account_disabled, not_a_server_user,
              password_sign_in_disabled, sign_in_method_unavailable, pin_expired,
              plex_pin_pending, plex_owner_required, plex_tv_unreachable,
              quick_connect_unavailable, quick_connect_expired,
              media_server_unreachable, media_server_changed;
            - pairing: pairing_expired, pairing_rejected,
              pairing_not_awaiting_approval, public_url_not_set,
              public_url_unverified;
            - connectors: connector_unreachable, connector_unauthorized,
              connector_unexpected_response, media_server_unsupported,
              media_server_required, secret_required;
            - swipe and admin: last_admin, streaming_region_not_set,
              metadata_unreachable, llm_not_configured, tmdb_not_configured,
              llm_auth_failed, llm_quota, llm_model_not_found, llm_unreachable,
              llm_invalid_output, daily_limit_reached, requests_not_configured,
              no_backend_user, title_not_offered, request_not_allowed,
              quota_exceeded, request_backend_error.
          examples: [llm_quota]
        retry_after_ms:
          type: integer
          minimum: 0
          description: Set on `rate_limited` and `daily_limit_reached`; the `Retry-After` header carries the same wait in seconds, rounded up.
        reason:
          type: string
          description: |
            Coarse cause for `public_url_unverified`, never a response body:
            `unreachable`, `tls_error`, `redirected`, `unexpected_response` or
            `proof_mismatch`.
          examples: [proof_mismatch]
        errors:
          type: array
          description: Field errors for `validation_error`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    Role:
      type: string
      enum: [admin, user]
    User:
      type: object
      required: [id, name, role]
      properties:
        id: { type: string }
        name: { type: string }
        role:
          $ref: "#/components/schemas/Role"
          description: Effective role, `admin` when `media_server_admin` or `promoted` (docs/adr/0010).
        avatar_url: { type: [string, "null"], format: uri }
        media_server_admin:
          type: boolean
          description: |
            Administrator on the media server, as read at the user's last sign-in
            (or cleared since). Only a media server administrator can change the
            media server connector, `public_url` and `password_sign_in`.
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [enabled, media_server_admin, promoted, remote_access, created_at, votes, generations_today]
          properties:
            enabled: { type: boolean }
            disabled_reason:
              type: [string, "null"]
              enum: [admin, media_server, unlinked, null]
              description: |
                Why the user is disabled: by an admin, by the media server (removed or
                disabled there; re-enabled at their next successful sign-in), or
                unlinked after the media server was replaced. Null when enabled.
            media_server_admin:
              type: boolean
              description: Administrator on the media server, as read at the user's last sign-in (or cleared by a demotion or the hourly sync since).
            promoted: { type: boolean, description: Made admin in Tindarr. }
            remote_access:
              type: boolean
              description: From the media server policy; when false, the user is refused from non-private client IPs.
            created_at: { type: string, format: date-time }
            last_sign_in_at: { type: [string, "null"], format: date-time }
            last_seen_at: { type: [string, "null"], format: date-time }
            votes: { type: integer, minimum: 0 }
            generations_today: { type: integer, minimum: 0 }
            daily_generation_limit: { type: [integer, "null"], minimum: 0 }
            request_backend_user_found: { type: boolean }
    Session:
      type: object
      required: [id, kind, device_name, platform, created_at, last_seen_at, current]
      properties:
        id: { type: string }
        kind: { type: string, enum: [mobile, web] }
        device_name:
          type: [string, "null"]
          description: The app's `DeviceInput.name`, or the browser a web session was opened from ("Firefox on Linux"). Null only for a row written without either.
        platform:
          type: [string, "null"]
          description: "`android`, `ios`, or `web` for a console session."
        app_version: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, description: Absolute end of the session. }
        current: { type: boolean }
...